+++ org.nibiru.ui.ios/src/main/java/org/nibiru/ui/ios/widget/IOSListWidget.java
@@ -4,12 +4,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import java.util.List;
 
-import org.nibiru.model.core.api.Type;
-import org.nibiru.model.core.api.Value;
-import org.nibiru.model.core.impl.BaseValue;
-import org.nibiru.model.core.impl.java.JavaType;
-import org.nibiru.ui.core.api.ListWidget;
-import org.nibiru.ui.core.api.Widget;
 import org.robovm.apple.coregraphics.CGRect;
 import org.robovm.apple.foundation.NSIndexPath;
 import org.robovm.apple.uikit.UITableView;
@@ -18,6 +12,12 @@ import org.robovm.apple.uikit.UITableViewDataSourceAdapter;
 import org.robovm.apple.uikit.UITableViewDelegateAdapter;
 import org.robovm.apple.uikit.UIView;
 
+import org.nibiru.model.core.api.Type;
+import org.nibiru.model.core.api.Value;
+import org.nibiru.model.core.impl.BaseValue;
+import org.nibiru.model.core.impl.java.JavaType;
+import org.nibiru.ui.core.api.ListWidget;
+import org.nibiru.ui.core.api.Widget;
 import com.google.common.collect.ImmutableList;
 
 public class IOSListWidget extends IOSValueWidget<UITableView, Iterable<Widget>> implements ListWidget {
@@ -81,7 +81,7 @@ public class IOSListWidget extends IOSValueWidget<UITableView, Iterable<Widget>>
 						width = childWidth;
 					}
 				}
-				updateSize(width, height);
+				updateSize((int)width, (int)height);
 
 			}
